// src/api/models/collection-preview.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UserPreview } from './user-preview';

/**
 * 
 * @export
 * @interface CollectionPreview
 */
export interface CollectionPreview {
    /**
     * 컬렉션 ID
     * @type {number}
     * @memberof CollectionPreview
     */
    'collectionId': number;
    /**
     * 컬렉션 제목
     * @type {string}
     * @memberof CollectionPreview
     */
    'title': string;
    /**
     * 컬렉션 설명
     * @type {string}
     * @memberof CollectionPreview
     */
    'description': string;
    /**
     * 
     * @type {UserPreview}
     * @memberof CollectionPreview
     */
    'author': UserPreview;
    /**
     * 컬렉션에 포함된 아이템 개수
     * @type {number}
     * @memberof CollectionPreview
     */
    'itemCount': number;
    /**
     * 받은 좋아요 수
     * @type {number}
     * @memberof CollectionPreview
     */
    'likeCount': number;
    /**
     * 커버 이미지 URL
     * @type {string}
     * @memberof CollectionPreview
     */
    'coverImageUrl': string;
    /**
     * 공개 여부
     * @type {string}
     * @memberof CollectionPreview
     */
    'visibility'?: CollectionPreviewVisibilityEnum;
}

export const CollectionPreviewVisibilityEnum = {
    Public: 'public',
    Private: 'private'
} as const;

export type CollectionPreviewVisibilityEnum = typeof CollectionPreviewVisibilityEnum[keyof typeof CollectionPreviewVisibilityEnum];