// src/api/models/get-my-activity200-response-all-of-data-user-preferences-inner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GetMyActivity200ResponseAllOfDataUserPreferencesInner
 */
export interface GetMyActivity200ResponseAllOfDataUserPreferencesInner {
    /**
     * 어떤 장르인지
     * @type {string}
     * @memberof GetMyActivity200ResponseAllOfDataUserPreferencesInner
     */
    'genre': string;
    /**
     * 몇점인지(최대 1)
     * @type {number}
     * @memberof GetMyActivity200ResponseAllOfDataUserPreferencesInner
     */
    'value': number;
}