// src/api/models/rated-album.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RatedAlbum
 */
export interface RatedAlbum {
    /**
     * 앨범 ID
     * @type {string}
     * @memberof RatedAlbum
     */
    'id': string;
    /**
     * 앨범 커버 이미지 URL
     * @type {string}
     * @memberof RatedAlbum
     */
    'imageUrl': string;
    /**
     * 앨범 제목
     * @type {string}
     * @memberof RatedAlbum
     */
    'title': string;
    /**
     * 아티스트 목록
     * @type {Array<string>}
     * @memberof RatedAlbum
     */
    'artists': Array<string>;
    /**
     * 발매일
     * @type {string}
     * @memberof RatedAlbum
     */
    'releaseDate'?: string;
    /**
     * 평균 평점 or 사용자가 준 평점
     * @type {number}
     * @memberof RatedAlbum
     */
    'rating': number;
    /**
     * 평가한 날짜
     * @type {string}
     * @memberof RatedAlbum
     */
    'ratedDate': string;
}