// src/api/models/rated-track.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Renew
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RatedTrack
 */
export interface RatedTrack {
    /**
     * 트랙 ID
     * @type {string}
     * @memberof RatedTrack
     */
    'id': string;
    /**
     * 트랙 제목
     * @type {string}
     * @memberof RatedTrack
     */
    'title': string;
    /**
     * 아티스트 목록
     * @type {Array<string>}
     * @memberof RatedTrack
     */
    'artists': Array<string>;
    /**
     * 앨범 커버 이미지 URL
     * @type {string}
     * @memberof RatedTrack
     */
    'imageUrl'?: string;
    /**
     * 발매일
     * @type {string}
     * @memberof RatedTrack
     */
    'releaseDate'?: string;
    /**
     * 평균 평점 or 사용자가 준 평점
     * @type {number}
     * @memberof RatedTrack
     */
    'rating': number;
    /**
     * 평가한 날짜
     * @type {string}
     * @memberof RatedTrack
     */
    'ratedDate': string;
}